requirements.txt
functions-framework==3.*
google-cloud-storage==2.14.*
